requirements.txt
joblib==1.0.1
matplotlib==3.8.0
numpy==1.19.2
omegaconf==2.3.0
onnx==1.14.1
Pillow==10.1.0
PyYAML==6.0.1
scikit_learn==1.3.1
seaborn==0.13.0
tensorflow==1.14.0
tf2onnx==1.15.1
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
tweepy==4.14.0
